Icewind Dale Ⅱ — Syntax update

+++ play.it-2/games/play-icewind-dale-2.sh
@@ -1,4 +1,4 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
@@ -30,42 +30,47 @@ set -o errexit
 
 ###
 # Icewind Dale 2
-# build native Linux packages from the original installers
-# send your bug reports to vv221@dotslashplay.it
+# build native packages from the original installers
+# send your bug reports to contact@dotslashplay.it
 ###
 
-script_version=20180806.1
+script_version=20200503.1
 
 # Set game-specific variables
 
 GAME_ID='icewind-dale-2'
-GAME_NAME='Icewind Dale II'
+GAME_NAME='Icewind Dale Ⅱ'
 
-ARCHIVES_LIST='ARCHIVE_GOG_EN ARCHIVE_GOG_FR'
+ARCHIVES_LIST='
+ARCHIVE_GOG_EN_0
+ARCHIVE_GOG_FR_0
+'
 
-ARCHIVE_GOG_EN='setup_icewind_dale2_2.1.0.13.exe'
-ARCHIVE_GOG_EN_URL='https://www.gog.com/game/icewind_dale_2'
-ARCHIVE_GOG_EN_MD5='9a68fdabdaff58bebc67092d47d4174e'
-ARCHIVE_GOG_EN_VERSION='2.01.101615-gog2.1.0.13'
-ARCHIVE_GOG_EN_SIZE='1500000'
+ARCHIVE_GOG_EN_0='setup_icewind_dale2_2.1.0.13.exe'
+ARCHIVE_GOG_EN_0_MD5='9a68fdabdaff58bebc67092d47d4174e'
+ARCHIVE_GOG_EN_0_TYPE='innosetup'
+ARCHIVE_GOG_EN_0_URL='https://www.gog.com/game/icewind_dale_2'
+ARCHIVE_GOG_EN_0_VERSION='2.01.101615-gog2.1.0.13'
+ARCHIVE_GOG_EN_0_SIZE='1500000'
 
-ARCHIVE_GOG_FR='setup_icewind_dale2_french_2.1.0.13.exe'
-ARCHIVE_GOG_FR_URL='https://www.gog.com/game/icewind_dale_2'
-ARCHIVE_GOG_FR_MD5='04f25433d405671a8975be6540dd55fa'
-ARCHIVE_GOG_FR_VERSION='2.01.101615-gog2.1.0.13'
-ARCHIVE_GOG_FR_SIZE='1500000'
+ARCHIVE_GOG_FR_0='setup_icewind_dale2_french_2.1.0.13.exe'
+ARCHIVE_GOG_FR_0_MD5='04f25433d405671a8975be6540dd55fa'
+ARCHIVE_GOG_FR_0_TYPE='innosetup'
+ARCHIVE_GOG_FR_0_URL='https://www.gog.com/game/icewind_dale_2'
+ARCHIVE_GOG_FR_0_VERSION='2.01.101615-gog2.1.0.13'
+ARCHIVE_GOG_FR_0_SIZE='1500000'
 
 ARCHIVE_DOC_L10N_PATH='app'
-ARCHIVE_DOC_L10N_FILES='./manual.pdf ./patch.txt ./readme.htm'
+ARCHIVE_DOC_L10N_FILES='manual.pdf patch.txt readme.htm'
 
 ARCHIVE_GAME_BIN_PATH='app'
-ARCHIVE_GAME_BIN_FILES='./binkw32.dll ./*.exe ./keymap.ini'
+ARCHIVE_GAME_BIN_FILES='binkw32.dll *.exe keymap.ini'
 
 ARCHIVE_GAME_L10N_PATH='app'
-ARCHIVE_GAME_L10N_FILES='./characters ./sounds ./*.tlk ./icewind2.ini ./language.ini ./party.ini ./override/narr002.wav ./override/guiinv.chu ./override/charstr.2da ./override/00bpak0.bcs ./data/guimos.bif ./cd2/data/intro.mve ./cd2/data/middle.mve ./cd2/data/sndvo.bif ./cd2/data/end.mve'
+ARCHIVE_GAME_L10N_FILES='characters sounds *.tlk icewind2.ini language.ini party.ini override/narr002.wav override/guiinv.chu override/charstr.2da override/00bpak0.bcs data/guimos.bif cd2/data/intro.mve cd2/data/middle.mve cd2/data/sndvo.bif cd2/data/end.mve'
 
 ARCHIVE_GAME_DATA_PATH='app'
-ARCHIVE_GAME_DATA_FILES='./cd2 ./chitin.key ./data ./music ./override ./scripts'
+ARCHIVE_GAME_DATA_FILES='cd2 chitin.key data music override scripts'
 
 CONFIG_FILES='./*.ini'
 DATA_DIRS='./characters ./mpsave ./override ./portraits ./scripts'
@@ -76,23 +81,24 @@ APP_WINETRICKS="vd=\$(xrandr|awk '/\\*/ {print \$1}')"
 APP_MAIN_TYPE='wine'
 APP_MAIN_EXE='iwd2.exe'
 APP_MAIN_ICON='iwd2.exe'
-APP_MAIN_ICON_RES='16 32 48'
 
 APP_CONFIG_ID="${GAME_ID}_config"
 APP_CONFIG_NAME="$GAME_NAME - configuration"
 APP_CONFIG_TYPE='wine'
 APP_CONFIG_EXE='config.exe'
 APP_CONFIG_ICON='config.exe'
-APP_CONFIG_ICON_RES='16 32 48'
 APP_CONFIG_CAT='Settings'
 
-PACKAGES_LIST='PKG_L10N PKG_DATA PKG_BIN'
+PACKAGES_LIST='PKG_BIN PKG_L10N PKG_DATA'
 
+# Common section for localization packages
 PKG_L10N_ID="${GAME_ID}-l10n"
-PKG_L10N_ID_GOG_EN="${PKG_L10N_ID}-en"
-PKG_L10N_ID_GOG_FR="${PKG_L10N_ID}-fr"
 PKG_L10N_PROVIDE="$PKG_L10N_ID"
+# English localization
+PKG_L10N_ID_GOG_EN="${PKG_L10N_ID}-en"
 PKG_L10N_DESCRIPTION_GOG_EN='English localization'
+# French localization
+PKG_L10N_ID_GOG_FR="${PKG_L10N_ID}-fr"
 PKG_L10N_DESCRIPTION_GOG_FR='French localization'
 
 PKG_DATA_ID="${GAME_ID}-data"
@@ -103,51 +109,62 @@ PKG_BIN_DEPS="$PKG_L10N_ID $PKG_DATA_ID wine winetricks xrandr"
 
 # Load common functions
 
-target_version='2.5'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
-tolower "$PLAYIT_WORKDIR/gamedata"
+prepare_package_layout
+rm --recursive "$PLAYIT_WORKDIR/gamedata"
 
-for PKG in $PACKAGES_LIST; do
-	organize_data "DOC_${PKG#PKG_}"  "$PATH_DOC"
-	organize_data "GAME_${PKG#PKG_}" "$PATH_GAME"
-done
+# Get game icons
 
 PKG='PKG_BIN'
-extract_and_sort_icons_from 'APP_MAIN' 'APP_CONFIG'
-move_icons_to 'PKG_DATA'
-
-rm --recursive "$PLAYIT_WORKDIR/gamedata"
+icons_get_from_package 'APP_MAIN' 'APP_CONFIG'
+icons_move_to 'PKG_DATA'
 
 # Update game path in .ini file
 
-file="${PKG_L10N_PATH}${PATH_GAME}/icewind2.ini"
-pattern="s/HD0:=.\\+/HD0:=C:\\\\${GAME_ID}\\\\/"
-pattern="$pattern;s/CD1:=.\\+/CD1:=C:\\\\${GAME_ID}\\\\data\\\\/"
-pattern="$pattern;s/CD2:=.\\+/CD2:=C:\\\\${GAME_ID}\\\\cd2\\\\/"
-sed --in-place "$pattern" "$file"
+pattern1='HD0:=.\+'
+replacement1='HD0:=C:\\'"${GAME_ID}"'\\'
+pattern2='CD1:=.\+'
+replacement2='CD1:=C:\\'"${GAME_ID}"'\\data\\'
+pattern3='CD2:=.\+'
+replacement3='CD2:=C:\\'"${GAME_ID}"'\\cd2\\'
+if [ $DRY_RUN -eq 0 ]; then
+	sed --in-place \
+		"s/${pattern1}/${replacement1}/;s/${pattern2}/${replacement2}/;s/${pattern3}/${replacement3}/" \
+		"${PKG_L10N_PATH}${PATH_GAME}/icewind2.ini"
+fi
 
 # Write launchers
 
 PKG='PKG_BIN'
-write_launcher 'APP_MAIN' 'APP_CONFIG'
+launchers_write 'APP_MAIN' 'APP_CONFIG'
 
 # Build package
